Move polygon methods into PolygonBuilder.ts
